refactor(tests): reorganize Task 1 tests into per-requirement layout
Move Task 1 test cases from tests/jobs/ and tests/integration/ into a single
per-task folder tests/01-polygon-area/, with one test file per Readme §1
requirement (R1, R2). Each file has a top-level describe naming the
requirement and inner happy/error describe blocks per CLAUDE.md §Tests.

All 10 existing tests are preserved (no coverage delta):

- outputs-area-in-square-meters.test.ts (R1: "output should include the
  calculated area in square meters") collects the JobFactory registration
  check, 1°x1° equator computation, Feature<Polygon> wrapper case, the
  workflow-drain happy-path integration, and the degenerate-ring edge case.
- handles-invalid-geojson-gracefully.test.ts (R2: "handles invalid GeoJSON
  gracefully and marks the task as failed") collects the malformed-JSON
  throw, the non-Polygon-type throw, the runner-level Result.error
  persistence test, and the worker-isolation integration scenario (US2,
  US20).

Fixture tests/test-workflows/polygonArea_only.yml moved to
tests/01-polygon-area/fixtures/polygonArea_only.yml (task-1-scoped). Old
files and now-empty parent dirs (tests/jobs/, tests/integration/,
tests/test-workflows/) removed. tests/smoke.test.ts is untouched.

Refs: CLAUDE.md §Tests, Readme.md §1 R1/R2, prior commit c6739e2.
